Move action changePassword from UserController to PasswordController #15
