+++ README.md
@@ -135,7 +135,7 @@ Route::group(['middleware'=>'auth'], function() { // use middleware jwt.auth if 
     Route::post('/auth/logout', '\PhpSoft\Users\Controllers\AuthController@logout');
     Route::get('/me', '\PhpSoft\Users\Controllers\UserController@authenticated');
     Route::patch('/me/profile', '\PhpSoft\Users\Controllers\UserController@updateProfile');
-    Route::put('/me/password', '\PhpSoft\Users\Controllers\UserController@changePassword');
+    Route::put('/me/password', '\PhpSoft\Users\Controllers\PasswordController@changePassword');
 });
 Route::post('/passwords/forgot', '\PhpSoft\Users\Controllers\PasswordController@forgot');
 Route::post('/passwords/reset', '\PhpSoft\Users\Controllers\PasswordController@reset');
