+++ app/Http/routes.php
@@ -21,7 +21,7 @@ Route::group(['middleware'=>'jwt.auth'], function() {
     Route::post('/auth/logout', '\PhpSoft\Users\Controllers\AuthController@logout');
     Route::get('/me', '\PhpSoft\Users\Controllers\UserController@authenticated');
     Route::patch('/me/profile', '\PhpSoft\Users\Controllers\UserController@updateProfile');
-    Route::put('/me/password', '\PhpSoft\Users\Controllers\UserController@changePassword');
+    Route::put('/me/password', '\PhpSoft\Users\Controllers\PasswordController@changePassword');
 });
 Route::post('/passwords/forgot', '\PhpSoft\Users\Controllers\PasswordController@forgot');
 Route::post('/passwords/reset', '\PhpSoft\Users\Controllers\PasswordController@reset');
