+++ packages/Users/Controllers/UserController.php
@@ -54,46 +54,6 @@ class UserController extends Controller
     }
 
     /**
-     * Change password
-     *
-     * @param  Request $request
-     * @return Response
-     */
-    public function changePassword(Request $request)
-    {
-        if (!$this->checkAuth()) {
-            return response()->json(null, 401);
-        }
-
-        $validator = Validator::make($request->all(), [
-            'old_password' => 'required|min:6',
-            'password'     => 'required|confirmed|min:6',
-        ]);
-
-        if ($validator->fails()) {
-            return response()->json(arrayView('phpsoft.users::errors/validation', [
-                'errors' => $validator->errors()
-            ]), 400);
-        }
-
-        $user = Auth::user();
-        $checkPassword = Auth::attempt(['id' => $user->id, 'password' => $request['old_password']]);
-        if (!$checkPassword) {
-            return response()->json(arrayView('phpsoft.users::errors/validation', [
-                'errors' => ['Old password is incorrect.']
-            ]), 401);
-        }
-
-        $change = $user->changePassword($request['password']);
-
-        if (!$change) {
-            return response()->json(null, 500); // @codeCoverageIgnore
-        }
-
-        return response()->json(null, 204);
-    }
-
-    /**
      * update profile action
      * @param  Request $request
      * @return Response
