+++ tests/Users/Controllers/PasswordControllerTest.php
@@ -85,4 +85,71 @@ class PasswordControllerTest extends TestCase
         ]);
         $this->assertEquals(200, $checkResetPassword->getStatusCode());
     }
+
+    public function testCheckAuthChangePassword()
+    {
+        $this->withoutMiddleware();
+        $res = $this->call('PUT', '/me/password');
+        $this->assertEquals(401, $res->getStatusCode());
+    }
+
+    public function testChangePassword()
+    {
+        // Check authenticate
+        $res = $this->call('PUT', '/me/password');
+        $results = json_decode($res->getContent());
+        $this->assertEquals('error', $results->status);
+        $this->assertEquals('authenticate', $results->type);
+        $this->assertEquals('Token is not provided.', $results->message);
+
+        $credentials = [ 'email' => 'admin@example.com', 'password' => '123456' ];
+        $token = JWTAuth::attempt($credentials);
+
+        // Input is empty
+        $res = $this->call('PUT', '/me/password', [], [], [], ['HTTP_Authorization' => "Bearer {$token}"]);
+        $this->assertEquals(400, $res->getStatusCode());
+        $results = json_decode($res->getContent());
+        $this->assertEquals('error', $results->status);
+        $this->assertEquals('validation', $results->type);
+        $this->assertObjectHasAttribute('old_password', $results->errors);
+        $this->assertEquals('The old password field is required.', $results->errors->old_password[0]);
+        $this->assertObjectHasAttribute('password', $results->errors);
+        $this->assertEquals('The password field is required.', $results->errors->password[0]);
+
+        // Check validate input
+        $res = $this->call('PUT', '/me/password', [
+            'old_password'          => '1234',
+            'password'              => '1234',
+            'password_confirmation' => '123'
+        ], [], [], ['HTTP_Authorization' => "Bearer {$token}"]);
+        $this->assertEquals(400, $res->getStatusCode());
+        $results = json_decode($res->getContent());
+        $this->assertEquals('error', $results->status);
+        $this->assertEquals('validation', $results->type);
+        $this->assertObjectHasAttribute('old_password', $results->errors);
+        $this->assertEquals('The old password must be at least 6 characters.', $results->errors->old_password[0]);
+        $this->assertObjectHasAttribute('password', $results->errors);
+        $this->assertEquals('The password confirmation does not match.', $results->errors->password[0]);
+        $this->assertEquals('The password must be at least 6 characters.', $results->errors->password[1]);
+
+        // Old password is wrong
+        $res = $this->call('PUT', '/me/password', [
+            'old_password'          => '123456789',
+            'password'              => '12345678',
+            'password_confirmation' => '12345678'
+        ], [], [], ['HTTP_Authorization' => "Bearer {$token}"]);
+        $this->assertEquals(401, $res->getStatusCode());
+        $results = json_decode($res->getContent());
+        $this->assertEquals('error', $results->status);
+        $this->assertEquals("Old password is incorrect.", $results->message);
+        $this->assertEquals('validation', $results->type);
+
+        // Change password success
+        $res = $this->call('PUT', '/me/password', [
+            'old_password'          => '123456',
+            'password'              => '12345678',
+            'password_confirmation' => '12345678'
+        ], [], [], ['HTTP_Authorization' => "Bearer {$token}"]);
+        $this->assertEquals(204, $res->getStatusCode());
+    }
 }
